[refactor] Refactor client.go
- Split types and clients into separate files
- Make inner clients private, outer client is now used for txns
- rename restclient -> nodeclient
