+++ client.go
@@ -1,19 +1,8 @@
 package aptos
 
 import (
-	"bytes"
-	"encoding/hex"
-	"encoding/json"
 	"errors"
-	"fmt"
-	"io"
-	"log/slog"
-	"net/http"
 	"net/url"
-	"os"
-	"path"
-	"strconv"
-	"strings"
 	"time"
 )
 
@@ -53,28 +42,22 @@ type NetworkConfig struct {
 	indexer *string
 }
 
+// Client is a facade over the multiple types of underlying clients, as the user doesn't actually care where the data
+// comes from.  It will be then handled underneath
 type Client struct {
-	// TODO: Move all these inner functions to the top level, then make these private
-	RestClient   NodeClient
-	FaucetClient FaucetClient
+	nodeClient   NodeClient
+	faucetClient FaucetClient
 	// TODO: Add indexer client
 }
 
-// TODO: rename 'NodeClient' (vs IndexerClient) ?
-// what looks best for `import aptos "github.com/aptoslabs/aptos-go-sdk"` then aptos.NewClient() ?
-type NodeClient struct {
-	ChainId uint8
-
-	client  http.Client
-	baseUrl url.URL
-}
-
+// NewClientFromNetworkName Creates a new client for a specific network name
 func NewClientFromNetworkName(network *string) (client *Client, err error) {
 	config := NetworkConfig{network: network}
 	client, err = NewClient(config)
 	return
 }
 
+// NewClient Creates a new client with a specific network config that can be extended in the future
 func NewClient(config NetworkConfig) (client *Client, err error) {
 	var apiUrl *url.URL = nil
 
@@ -160,217 +143,29 @@ func NewClient(config NetworkConfig) (client *Client, err error) {
 	return
 }
 
-type NodeInfo struct {
-	ChainId                uint8  `json:"chain_id"`
-	EpochStr               string `json:"epoch"`
-	LedgerVersionStr       string `json:"ledger_version"`
-	OldestLedgerVersionStr string `json:"oldest_ledger_version"`
-	NodeRole               string `json:"node_role"`
-	BlockHeightStr         string `json:"block_height"`
-	OldestBlockHeightStr   string `json:"oldest_block_height"`
-	GitHash                string `json:"git_hash"`
+func (client *Client) Info() (info NodeInfo, err error) {
+	return client.nodeClient.Info()
 }
 
-// TODO: write NodeInfo accessors to ParseUint on *Str which work around 53 bit float64 limit in JavaScript
-
-func (rc *NodeClient) Info() (info NodeInfo, err error) {
-	response, err := rc.client.Get(rc.baseUrl.String())
-	if err != nil {
-		err = fmt.Errorf("GET %s, %w", rc.baseUrl.String(), err)
-		return
-	}
-	if response.StatusCode >= 400 {
-		err = NewHttpError(response)
-		return
-	}
-	blob, err := io.ReadAll(response.Body)
-	if err != nil {
-		err = fmt.Errorf("error getting response data, %w", err)
-		return
-	}
-	_ = response.Body.Close()
-	err = json.Unmarshal(blob, &info)
-	if err == nil {
-		rc.ChainId = info.ChainId
-	}
-	return
+func (client *Client) Account(address AccountAddress, ledgerVersion ...int) (info AccountInfo, err error) {
+	return client.nodeClient.Account(address, ledgerVersion...)
 }
 
 // TODO: set HTTP header "x-aptos-client: aptos-go-sdk/{version}"
 
-// AccountInfo is returned from calls to #Account()
-type AccountInfo struct {
-	SequenceNumberStr    string `json:"sequence_number"`
-	AuthenticationKeyHex string `json:"authentication_key"`
-}
-
-// Hex decode of AuthenticationKeyHex
-func (ai AccountInfo) AuthenticationKey() ([]byte, error) {
-	ak := ai.AuthenticationKeyHex
-	if strings.HasPrefix(ak, "0x") {
-		ak = ak[2:]
-	}
-	return hex.DecodeString(ak)
-}
-
-// ParseUint of SequenceNumberStr
-func (ai AccountInfo) SequenceNumber() (uint64, error) {
-	return strconv.ParseUint(ai.SequenceNumberStr, 10, 64)
-}
-
-func (rc *NodeClient) Account(address AccountAddress, ledger_version ...int) (info AccountInfo, err error) {
-	au := rc.baseUrl
-	au.Path = path.Join(au.Path, "accounts", address.String())
-	if len(ledger_version) > 0 {
-		params := url.Values{}
-		params.Set("ledger_version", strconv.Itoa(ledger_version[0]))
-		au.RawQuery = params.Encode()
-	}
-	response, err := rc.client.Get(au.String())
-	if err != nil {
-		err = fmt.Errorf("GET %s, %w", au.String(), err)
-		return
-	}
-	if response.StatusCode >= 400 {
-		err = NewHttpError(response)
-		return
-	}
-	blob, err := io.ReadAll(response.Body)
-	if err != nil {
-		err = fmt.Errorf("error getting response data, %w", err)
-		return
-	}
-	_ = response.Body.Close()
-	err = json.Unmarshal(blob, &info)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "account json err: %v\n%s\n", err, string(blob))
-	}
-	return
-}
-
-// AccountResourceInfo is returned by #AccountResource() and #AccountResources()
-type AccountResourceInfo struct {
-	// e.g. "0x1::coin::CoinStore<0x1::aptos_coin::AptosCoin>"
-	Type string `json:"type"`
-
-	// Decoded from Move contract data, could really be anything
-	Data map[string]any `json:"data"`
-}
-
-func (rc *NodeClient) AccountResource(address AccountAddress, resourceType string, ledger_version ...int) (data map[string]any, err error) {
-	au := rc.baseUrl
-	// TODO: offer a list of known-good resourceType string constants
-	// TODO: set "Accept: application/x-bcs" and parse BCS objects for lossless (and faster) transmission
-	au.Path = path.Join(au.Path, "accounts", address.String(), "resource", resourceType)
-	if len(ledger_version) > 0 {
-		params := url.Values{}
-		params.Set("ledger_version", strconv.Itoa(ledger_version[0]))
-		au.RawQuery = params.Encode()
-	}
-	response, err := rc.client.Get(au.String())
-	if err != nil {
-		err = fmt.Errorf("GET %s, %w", au.String(), err)
-		return
-	}
-	if response.StatusCode >= 400 {
-		err = NewHttpError(response)
-		return
-	}
-	blob, err := io.ReadAll(response.Body)
-	if err != nil {
-		err = fmt.Errorf("error getting response data, %w", err)
-		return
-	}
-	_ = response.Body.Close()
-	err = json.Unmarshal(blob, &data)
-	return
+func (client *Client) AccountResource(address AccountAddress, resourceType string, ledgerVersion ...int) (data map[string]any, err error) {
+	return client.nodeClient.AccountResource(address, resourceType, ledgerVersion...)
 }
 
 // AccountResources fetches resources for an account into a JSON-like map[string]any in AccountResourceInfo.Data
 // For fetching raw Move structs as BCS, See #AccountResourcesBCS
-func (rc *NodeClient) AccountResources(address AccountAddress, ledger_version ...int) (resources []AccountResourceInfo, err error) {
-	au := rc.baseUrl
-	au.Path = path.Join(au.Path, "accounts", address.String(), "resources")
-	if len(ledger_version) > 0 {
-		params := url.Values{}
-		params.Set("ledger_version", strconv.Itoa(ledger_version[0]))
-		au.RawQuery = params.Encode()
-	}
-	response, err := rc.client.Get(au.String())
-	if err != nil {
-		err = fmt.Errorf("GET %s, %w", au.String(), err)
-		return
-	}
-	if response.StatusCode >= 400 {
-		err = NewHttpError(response)
-		return
-	}
-	blob, err := io.ReadAll(response.Body)
-	if err != nil {
-		err = fmt.Errorf("error getting response data, %w", err)
-		return
-	}
-	_ = response.Body.Close()
-	err = json.Unmarshal(blob, &resources)
-	return
-}
-
-// DeserializeSequence[AccountResourceRecord](bcs) approximates the Rust side BTreeMap<StructTag,Vec<u8>>
-// They should BCS the same with a prefix Uleb128 length followed by (StructTag,[]byte) pairs.
-type AccountResourceRecord struct {
-	// Account::Module::Name
-	Tag StructTag
-
-	// BCS data as stored by Move contract
-	Data []byte
-}
-
-func (aar *AccountResourceRecord) MarshalBCS(bcs *Serializer) {
-	aar.Tag.MarshalBCS(bcs)
-	bcs.WriteBytes(aar.Data)
-}
-func (aar *AccountResourceRecord) UnmarshalBCS(bcs *Deserializer) {
-	aar.Tag.UnmarshalBCS(bcs)
-	aar.Data = bcs.ReadBytes()
-}
-
-func (rc *NodeClient) GetBCS(getUrl string) (*http.Response, error) {
-	req, err := http.NewRequest("GET", getUrl, nil)
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Set("Accept", "application/x-bcs")
-	return rc.client.Do(req)
+func (client *Client) AccountResources(address AccountAddress, ledgerVersion ...int) (resources []AccountResourceInfo, err error) {
+	return client.nodeClient.AccountResources(address, ledgerVersion...)
 }
 
 // AccountResourcesBCS fetches account resources as raw Move struct BCS blobs in AccountResourceRecord.Data []byte
-func (rc *NodeClient) AccountResourcesBCS(address AccountAddress, ledger_version ...int) (resources []AccountResourceRecord, err error) {
-	au := rc.baseUrl
-	au.Path = path.Join(au.Path, "accounts", address.String(), "resources")
-	if len(ledger_version) > 0 {
-		params := url.Values{}
-		params.Set("ledger_version", strconv.Itoa(ledger_version[0]))
-		au.RawQuery = params.Encode()
-	}
-	response, err := rc.GetBCS(au.String())
-	if err != nil {
-		err = fmt.Errorf("GET %s, %w", au.String(), err)
-		return
-	}
-	if response.StatusCode >= 400 {
-		err = NewHttpError(response)
-		return
-	}
-	blob, err := io.ReadAll(response.Body)
-	if err != nil {
-		err = fmt.Errorf("error getting response data, %w", err)
-		return
-	}
-	response.Body.Close()
-	bcs := NewDeserializer(blob)
-	// See resource_test.go TestMoveResourceBCS
-	resources = DeserializeSequence[AccountResourceRecord](bcs)
-	return
+func (client *Client) AccountResourcesBCS(address AccountAddress, ledgerVersion ...int) (resources []AccountResourceRecord, err error) {
+	return client.nodeClient.AccountResourcesBCS(address, ledgerVersion...)
 }
 
 // TransactionByHash gets info on a transaction
@@ -388,203 +183,35 @@ func (rc *NodeClient) AccountResourcesBCS(address AccountAddress, ledger_version
 //			// known to local mempool, but not committed yet
 //		}
 //	}
-func (rc *NodeClient) TransactionByHash(txnHash string) (data map[string]any, err error) {
-	restUrl := rc.baseUrl
-	restUrl.Path = path.Join(restUrl.Path, "transactions/by_hash", txnHash)
-	return rc.getTransactionCommon(restUrl)
+func (client *Client) TransactionByHash(txnHash string) (data map[string]any, err error) {
+	return client.nodeClient.TransactionByHash(txnHash)
 }
 
-func (rc *NodeClient) TransactionByVersion(version uint64) (data map[string]any, err error) {
-	restUrl := rc.baseUrl
-	restUrl.Path = path.Join(restUrl.Path, "transactions/by_version", strconv.FormatUint(version, 10))
-	return rc.getTransactionCommon(restUrl)
+func (client *Client) TransactionByVersion(version uint64) (data map[string]any, err error) {
+	return client.nodeClient.TransactionByVersion(version)
 }
 
-func (rc *NodeClient) getTransactionCommon(restUrl url.URL) (data map[string]any, err error) {
-	// Fetch transaction
-	response, err := rc.client.Get(restUrl.String())
-	if err != nil {
-		err = fmt.Errorf("GET %s, %w", restUrl.String(), err)
-		return
-	}
-
-	// Handle Errors TODO: Handle ratelimits, etc.
-	if response.StatusCode >= 400 {
-		err = NewHttpError(response)
-		return
-	}
-
-	// Read body to JSON TODO: BCS
-	blob, err := io.ReadAll(response.Body)
-	if err != nil {
-		err = fmt.Errorf("error getting response data, %w", err)
-		return
-	}
-	_ = response.Body.Close() // We don't care about the error about closing the body
-	err = json.Unmarshal(blob, &data)
-	return
-}
-
-// Waits up to 10 seconds for transactions to be done, polling at 10Hz
+// WaitForTransactions Waits up to 10 seconds for transactions to be done, polling at 10Hz
 // TODO: options for polling period and timeout
-func (rc *NodeClient) WaitForTransactions(txnHashes []string) error {
-	hashSet := make(map[string]bool, len(txnHashes))
-	for _, hash := range txnHashes {
-		hashSet[hash] = true
-	}
-	start := time.Now()
-	deadline := start.Add(10 * time.Second)
-	for len(hashSet) > 0 {
-		if time.Now().After(deadline) {
-			return errors.New("timeout waiting for faucet transactions")
-		}
-		time.Sleep(100 * time.Millisecond)
-		for _, hash := range txnHashes {
-			if !hashSet[hash] {
-				// already done
-				continue
-			}
-			status, err := rc.TransactionByHash(hash)
-			if err == nil {
-				if status["type"] == "pending_transaction" {
-					// not done yet!
-				} else if truthy(status["success"]) {
-					// done!
-					delete(hashSet, hash)
-					slog.Debug("txn done", "hash", hash, "status", status["success"])
-				}
-			}
-		}
-	}
-	return nil
+func (client *Client) WaitForTransactions(txnHashes []string) error {
+	return client.nodeClient.WaitForTransactions(txnHashes)
 }
 
-// Get recent transactions.
+// Transactions Get recent transactions.
 // Start is a version number. Nil for most recent transactions.
 // Limit is a number of transactions to return. 'about a hundred' by default.
-func (rc *NodeClient) Transactions(start *uint64, limit *uint64) (data []map[string]any, err error) {
-	au := rc.baseUrl
-	au.Path = path.Join(au.Path, "transactions")
-	var params url.Values
-	if start != nil {
-		params.Set("start", strconv.FormatUint(*start, 10))
-	}
-	if limit != nil {
-		params.Set("limit", strconv.FormatUint(*limit, 10))
-	}
-	if len(params) != 0 {
-		au.RawQuery = params.Encode()
-	}
-	// TODO: ?limit=N&start=V
-	response, err := rc.client.Get(au.String())
-	if err != nil {
-		err = fmt.Errorf("GET %s, %w", au.String(), err)
-		return
-	}
-	if response.StatusCode >= 400 {
-		err = NewHttpError(response)
-		return
-	}
-	blob, err := io.ReadAll(response.Body)
-	if err != nil {
-		err = fmt.Errorf("error getting response data, %w", err)
-		return
-	}
-	_ = response.Body.Close()
-	err = json.Unmarshal(blob, &data)
-	return
+func (client *Client) Transactions(start *uint64, limit *uint64) (data []map[string]any, err error) {
+	return client.nodeClient.Transactions(start, limit)
 }
 
-// testing only
-// There exists an aptos-node API for submitting JSON and having the node Rust code encode it to BCS, we should only use this for testing to validate our local BCS. Actual GO-SDK usage should use BCS encoding locally in Go code.
-func (rc *NodeClient) transactionEncode(request map[string]any) (data []byte, err error) {
-	rblob, err := json.Marshal(request)
-	if err != nil {
-		return
-	}
-	bodyReader := bytes.NewReader(rblob)
-	au := rc.baseUrl
-	au.Path = path.Join(au.Path, "transactions/encode_submission")
-	response, err := rc.client.Post(au.String(), "application/json", bodyReader)
-	if err != nil {
-		err = fmt.Errorf("POST %s, %w", au.String(), err)
-		return
-	}
-	if response.StatusCode >= 400 {
-		err = NewHttpError(response)
-		return
-	}
-	blob, err := io.ReadAll(response.Body)
-	if err != nil {
-		err = fmt.Errorf("error getting response data, %w", err)
-		return
-	}
-	_ = response.Body.Close()
-	err = json.Unmarshal(blob, &data)
-	return
+func (client *Client) SubmitTransaction(signedTransaction *SignedTransaction) (data map[string]any, err error) {
+	return client.nodeClient.SubmitTransaction(signedTransaction)
 }
 
-func (rc *NodeClient) SubmitTransaction(stxn *SignedTransaction) (data map[string]any, err error) {
-	bcs := Serializer{}
-	stxn.MarshalBCS(&bcs)
-	err = bcs.Error()
-	if err != nil {
-		return
-	}
-	sblob := bcs.ToBytes()
-	bodyReader := bytes.NewReader(sblob)
-	au := rc.baseUrl
-	au.Path = path.Join(au.Path, "transactions")
-	response, err := rc.client.Post(au.String(), APTOS_SIGNED_BCS, bodyReader)
-	if err != nil {
-		err = fmt.Errorf("POST %s, %w", au.String(), err)
-		return
-	}
-	if response.StatusCode >= 400 {
-		err = NewHttpError(response)
-		return nil, err
-	}
-	blob, err := io.ReadAll(response.Body)
-	if err != nil {
-		err = fmt.Errorf("error getting response data, %w", err)
-		return
-	}
-	_ = response.Body.Close()
-	//return blob, nil
-	err = json.Unmarshal(blob, &data)
-	return
+func (client *Client) GetChainId() (chainId uint8, err error) {
+	return client.nodeClient.GetChainId()
 }
 
-func (rc *NodeClient) GetChainId() (chainId uint8, err error) {
-	if rc.ChainId != 0 {
-		return rc.ChainId, nil
-	}
-	info, err := rc.Info()
-	if err != nil {
-		return 0, err
-	}
-	return info.ChainId, nil
-}
-
-type HttpError struct {
-	Status     string // e.g. "200 OK"
-	StatusCode int    // e.g. 200
-	Header     http.Header
-	Body       []byte
-}
-
-func NewHttpError(response *http.Response) *HttpError {
-	body, _ := io.ReadAll(response.Body)
-	_ = response.Body.Close()
-	return &HttpError{
-		Status:     response.Status,
-		StatusCode: response.StatusCode,
-		Header:     response.Header,
-		Body:       body,
-	}
-}
-
-// implement error interface
-func (he *HttpError) Error() string {
-	return fmt.Sprintf("HttpError %#v (%d bytes %s)", he.Status, len(he.Body), he.Header.Get("Content-Type"))
+func (client *Client) Fund(address AccountAddress, amount uint64) error {
+	return client.faucetClient.Fund(address, amount)
 }
